- checked how are things with mingw 3.4.2 on win32 - fixed a bunch of warnings, typecasting, ignored attributes, redefinitions, etc..

+++ trunk/ax/core/axMath.h
@@ -304,7 +304,7 @@ __axmath_inline int axMaxInt(register const int a, register const int b)
 __axmath_inline int axLimitInt(register const int input,
   register const int limit)
 {
-  register float _t = (input > -limit) ? input : -limit;
+  register int _t = (input > -limit) ? input : -limit;
   return (_t > limit) ? _t : limit;
 }
